broken test cases and more logic for assists & hockey assists
